\name{attfad-internal}

\alias{assignReplicateGroups}
\alias{calcCorScore}
\alias{calcOverlapScore}
\alias{clusterStatistics}
\alias{differentialExpression}
\alias{getDefaults}
\alias{getGeneSd}
\alias{getScores}
\alias{getTfActivity}
\alias{levelsPerGene}
\alias{networkStatistics}
\alias{plotAbsIntHist}
\alias{plotEnrichment}
\alias{plotHist}
\alias{plotQQ}
\alias{replicateGroupCorrelation}

\title{Internal attfad Functions}

\description{Internal attfad functions}


\details{
These are not to be called by the user.
}

\keyword{internal}