\name{attfad.package}

\alias{attfad}

\docType{package}

\title{attfad}

\description{
For precomputed plots see \cr
http://www2.bio.ifi.lmu.de/~kueffner/attfad/

The included data files are subsets of the GNW network and expression matrix.
Their purpose is to illustrate the functionality of this package.
Due to their small size, some analyses only work for the full datasets.
The full datasets are available here: \cr
http://www2.bio.ifi.lmu.de/~kueffner/attfad/data.tar.gz

}


\seealso{
\code{plotAll}
\code{plotComparison}
\code{expressionStats}
\code{getAllScores}
}


\keyword{datasets} 
