\name{GNW.s.exp.group}

\alias{GNW.s.exp.group}

\docType{data}

\title{Replicate groups for experiments in 'GNW.s.expression.data'}
\description{
  This is used to indicate to which replicate group an experiment belongs.
}
\usage{GNW.s.exp.group}
\format{A character vector of length 100.}

\keyword{datasets} 
