\name{GNW.s.expression.data}

\alias{GNW.s.expression.data}

\docType{data}

\title{Subset of an expression data matrix from GeneNetWeaver}
\description{
  This matrix contains expression values for 100 experiments and 500 genes
  which were simulated by GeneNetWeaver.
}
\usage{GNW.s.expression.data}
\format{A matrix with 100 rows corresponding to 100 experiments and 500 columns corresponding to 500 genes.}

\keyword{datasets} 
