\name{GNW.s.network}

\alias{GNW.s.network}

\docType{data}

\title{Subset of a gene regulatory network from GeneNetWeaver}
\description{
  A gene regulatory network used by GeneNetWeaver for simulation of expression data.
  The network is represented as an edge list with transcription factors in the first column and target genes in the second column.
}
\usage{GNW.s.network}
\format{A two column character matrix with one interaction per row.}

\keyword{datasets} 
